fix link to the portworx README
Markdown links don't generally require the `.md` extension when viewed in generated documentation, but in generated GitHub links the `.md` extension is needed. 

There are conflicting interests in how to represent these links. What works on GitHub won't work in generated docs and neither of those approaches work on the Terraform Registry: 
* modules/portworx/README
* modules/portworx/README.md
* https://github.com/equinix/terraform-metal-anthos-on-baremetal/blob/master/modules/portworx/README.md
* https://registry.terraform.io/modules/equinix/anthos-on-baremetal/metal/latest/submodules/portworx.

This PR includes two links, the former should keep GitHub users on GitHub, the latter will work on the Terraform Registry.